+++ CHANGELOG.md
@@ -6,6 +6,9 @@
   - inherits from ActiveRecord::RecordInvalid so rescuing is not affected
   - makes it more obvious where the error is coming from
 
+### Fixed
+- fixed NoMethodError when updating related errors that have errors on other, non-denormalized fields
+
 ## v1.2.1
 ### Fixed
 - relaxed dependency spec to include rails 7